+++ Program/Server/src/Domain/Interfaces/Repositories/RoomRepository/IRoomRepository.cs
@@ -3,7 +3,7 @@ using Domain.Interfaces.Repositories.BaseRepository;
 
 namespace Domain.Interfaces.Repositories.RoomRepository;
 
-public interface IRoomRepository : IBaseCrudRepository<Room, ushort>
+public interface IRoomRepository : IBaseCrudRepository<Room, int>
 {
 
 }
